requirements.txt
streamlit
openai
faiss-cpu
sentence-transformers
pandas
python-docx
python-pptx
PyPDF2